moves suggestions to all pieces, dynamic filters, get only words
