+++ lib/modules/apostrophe-images/index.js
@@ -162,7 +162,6 @@ module.exports = {
     self.pushAsset('script', 'focal-point-editor', { when: 'user' });
     self.pushAsset('stylesheet', 'user', { when: 'user' });
     require('./lib/api.js')(self, options);
-    require('./lib/routes.js')(self, options);
     self.enableHelpers();
   }
 };
