+++ lib/modules/apostrophe-images/lib/api.js
@@ -150,45 +150,4 @@ module.exports = function(self, options) {
     projection.attachment = 1;
     return projection;
   };
-
-  // This method gets titles and tags suggestions based on
-  // the autocomplete cursor, it also checks for filters.
-  self.getSuggestions = async function(req, term, filters) {
-    const search = self.apos.launder.string(term).toLowerCase();
-    const { published, trash, orientation, fileType, mediaSource } = filters;
-    const publishedVal = getFilterValue(published);
-    const trashVal = getFilterValue(trash);
-
-    const images = await self.find(req, {})
-      .published(publishedVal)
-      .trash(trashVal)
-      .orientation(orientation)
-      .fileType(fileType)
-      .mediaSource(mediaSource)
-      .autocomplete(search).limit(15).toArray();
-
-    const suggestions = images.reduce((acc, img) => {
-      if (acc.length >= 10) {
-        return acc;
-      }
-
-      const title = img.title.toLowerCase();
-      const titleMerge = (title.includes(search) && !acc.includes(title)) ? [ img.title ] : [];
-      const tagsMerge = img.tags.filter((tag) => !acc.includes(tag) && tag.startsWith(search));
-
-      return [
-        ...acc,
-        ...titleMerge,
-        ...tagsMerge
-      ];
-    }, []);
-
-    return suggestions;
-
-    function getFilterValue (value) {
-      return typeof value === 'boolean'
-        ? value
-        : typeof value === 'string' ? parseInt(value, 10) : null;
-    }
-  };
 };
