+++ lib/modules/apostrophe-pieces/lib/api.js
@@ -124,6 +124,45 @@ module.exports = function(self, options) {
     return next();
   };
 
+  // This method gets titles and tags suggestions based on
+  // the autocomplete cursor, it also checks for filters.
+  self.getSuggestions = async function(req, term, filters) {
+    const search = self.apos.launder.string(term).toLowerCase();
+    const cursor = getCursorWithFilters(filters);
+    const images = await cursor.autocomplete(search).limit(15).toArray();
+
+    const suggestions = images.reduce((acc, img) => {
+      if (acc.length >= 10) {
+        return acc;
+      }
+
+      const matchingWords = img.highSearchWords.filter((word) => {
+        if (acc.includes(word)) {
+          return false;
+        }
+        return acc.includes(word)
+          ? false
+          : word.startsWith(search);
+      });
+
+      return [
+        ...acc,
+        ...matchingWords
+      ];
+    }, []);
+
+    return suggestions;
+
+    // adding filter to cursor only if this one exist for the current piece
+    function getCursorWithFilters (filters) {
+      return Object.entries(filters).reduce((cursor, [filter, value]) => {
+        return cursor[filter]
+          ? cursor[filter](['0', '1'].includes(value) ? parseInt(value, 10) : value)
+          : cursor;
+      }, self.find(req, {}));
+    };
+  };
+
   // options.filters can contain cursor filters. `options.chooser`, `options.format` and
   // `options.manageView` are also implemented. For bc, if `options.filters` does not exist,
   // all properties of options are treated as cursor filters.
