+++ lib/modules/apostrophe-pieces/lib/routes.js
@@ -7,6 +7,7 @@ module.exports = function(self, options) {
 
     self.route('post', 'insert', self.requireEditor, self.routes.insert);
     self.route('post', 'retrieve', self.requirePieceEditorView, self.routes.retrieve);
+    self.route('post', 'suggestions', self.apos.docs.requireEditorOfSomething, self.routes.suggestions);
     self.route('post', 'list', self.apos.docs.requireEditorOfSomething, self.routes.list);
     self.route('post', 'update', self.requirePiece, self.routes.update);
     self.route('post', 'publish', self.requireEditor, self.routes.publish);
@@ -32,6 +33,28 @@ module.exports = function(self, options) {
 
   self.routes.retrieve = function(req, res) {
     return self.retrieveResponse(req, res, null, req.piece);
+  };
+
+  self.routes.suggestions = async function(req, res) {
+    try {
+      const {term, filters} = req.body;
+
+      if (!term || typeof filters !== 'object') {
+        const errorMsg = !term
+          ? 'A search term must be provided.'
+          : 'Filters must be an object containing search filters.';
+
+        return self.apiResponse(res, errorMsg);
+        // return res.status(400).send(msg);
+      }
+
+      const suggestions = await self.getSuggestions(req, term, filters);
+
+      return self.apiResponse(res, null, suggestions);
+
+    } catch (err) {
+      return self.apiResponse(res, err);
+    }
   };
 
   self.routes.list = function(req, res) {
