+++ lib/modules/apostrophe-pieces/public/js/manager-modal.js
@@ -619,18 +619,141 @@ apos.define('apostrophe-pieces-manager-modal', {
       });
     };
 
+    // self.enableSearch = function() {
+    //   self.$filters.on('change', '[name="search-' + self.options.name + '"]', function() {
+    //     self.search = $(this).val();
+    //     self.currentFilters.page = 1;
+    //     self.refresh(function() {
+    //       var $input = self.$filters.find('[name="search-' + self.options.name + '"]');
+    //       // refocus input element after search
+    //       $input.focus();
+    //       // put the search query before the cursor
+    //       $input.val($input.val());
+    //     });
+    //   });
+    // };
+
     self.enableSearch = function() {
-      self.$filters.on('change', '[name="search-' + self.options.name + '"]', function() {
-        self.search = $(this).val();
-        self.currentFilters.page = 1;
-        self.refresh(function() {
-          var $input = self.$filters.find('[name="search-' + self.options.name + '"]');
-          // refocus input element after search
-          $input.focus();
-          // put the search query before the cursor
-          $input.val($input.val());
+      var isListening = false;
+      var inputSelector = '[name="search-' + self.options.name + '"]';
+
+      self.$filters.on('keyup', inputSelector, function(e) {
+        var $input = $(e.target);
+        var $search = self.$filters.find('.apos-modal-filters-search');
+        var $suggestionsList = $search.find('.ui-autocomplete');
+        var $activeItem = $suggestionsList.find('li.ui-menu-item.active');
+        var listener = getListener($search);
+        var inputValue = $input.val();
+
+        if (e.keyCode === 13) {
+          return search($input, $activeItem);
+        }
+
+        if (!inputValue.length) {
+          return $search.find('ul').remove();
+        }
+
+        // If up or down arrows, we select suggestions in list
+        if ([38, 40].includes(e.keyCode)) {
+          return selectItems($suggestionsList, $activeItem, e.keyCode);
+        }
+
+        // TODO: search only or current word
+
+        var options = {
+          term: inputValue,
+          filters: self.currentFilters
+        };
+
+        self.api('suggestions', options, function(results) {
+          if (results.status === 'ok' && results.data.length) {
+            var suggestions = '<ul class="ui-autocomplete">';
+            results.data.forEach(function (result, i) {
+              suggestions = suggestions.concat('<li class="ui-menu-item">' + result + '</li>');
+              if (i === results.data.length - 1) {
+                suggestions = suggestions.concat('</ul>');
+              }
+            });
+
+            $search.find('ul').remove();
+            $search.append(suggestions);
+
+            $search.find('ul > .ui-menu-item').each(function (i, elem) {
+              var $el = $(elem);
+
+              $(elem).on('click', function() {
+                search($input, $el);
+              });
+            });
+
+            if (!isListening) {
+              isListening = true;
+              window.addEventListener('click', listener);
+            }
+
+          } else {
+            if (isListening) {
+              listener();
+            }
+          }
         });
       });
+
+      function search ($input, $activeItem) {
+        var val;
+
+        if ($activeItem && $activeItem.length) {
+          val = $activeItem.text();
+        } else {
+          val = $input.val();
+        }
+
+        self.search = val;
+        self.currentFilters.page = 1;
+
+        self.refresh(function() {
+          var $refreshedInput = self.$filters.find('[name="search-' + self.options.name + '"]');
+
+          $refreshedInput.focus();
+        });
+      };
+
+      function getListener ($search) {
+        return function listener () {
+          $search.find('ul').remove();
+          isListening = false;
+          window.removeEventListener('click', listener);
+        };
+      };
+
+      function selectItems ($suggestionsList, $activeItem, keyCode) {
+        if (!$suggestionsList.length) {
+          return;
+        }
+
+        var isDown = keyCode === 40;
+        var firstOrLastSelector = ':last-child';
+        if (isDown) {
+          firstOrLastSelector = ':first-child';
+        }
+
+        if ($activeItem.length) {
+          var $nextItem;
+          if (isDown) {
+            $nextItem = $activeItem.next();
+          } else {
+            $nextItem = $activeItem.prev();
+          }
+
+          if ($nextItem.length) {
+            $nextItem.focus().addClass('active');
+          }
+          $activeItem.removeClass('active');
+
+        } else {
+          $suggestionsList.find('li' + firstOrLastSelector).focus().addClass('active');
+        }
+      }
     };
 
     self.choices = [];
